Working db connection and graph view

+++ src/main/java/com/example/f21comp1011assignment1/Controllers/GraphViewController.java
@@ -19,7 +19,7 @@ import javafx.scene.control.ComboBox;
 public class GraphViewController implements Initializable{
 
     @FXML
-    private BarChart<String, Float> barChart;
+    private BarChart<String, Double> barChart;
 
     @FXML
     private ComboBox<String> nutrientComboBox;
